refactor emr
