+++ src/app.module.ts
@@ -14,7 +14,7 @@ import { TransactionLoggingModule } from './transaction-logging/transaction-logg
 import { TransactionRequest } from './transaction-logging/models/transaction-request.entity';
 import { RecaptchaModule } from './recaptcha/recaptcha.module';
 import { MailModule } from './mail/mail.module';
-import { EmrModule } from './emr/emr.module';
+import { EmrModule } from './category/emr/emr.module';
 // import dotenv from 'dotenv';
 
 require('dotenv').config(); // eslint-disable-line
